+++ src/bin/bat/clap_app.rs
@@ -477,6 +477,13 @@ pub fn build_app(interactive_output: bool) -> ClapApp<'static, 'static> {
                 .hidden(true)
                 .help("Show bat's cache directory."),
         )
+        .arg(
+            Arg::with_name("diagnostic")
+                .long("diagnostic")
+                .alias("diagnostics")
+                .hidden_short_help(true)
+                .help("Show diagnostic information for bug reports.")
+        )
         .help_message("Print this help message.")
         .version_message("Show version information.");